requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
fastapi-pagination          # ➕
fastapi-cache2              # ➕
slowapi                     # ➕
prometheus-fastapi-instrumentator  # ➕
redis>=4.6
